% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delete_users_permissions.R
\name{delete_user_from_account}
\alias{delete_user_from_account}
\title{function to delete user from account}
\usage{
delete_user_from_account(acc_id, email_to_delete)
}
\description{
function to delete user from account
}
