% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delete_users_permissions.R
\name{delete_user_from_view}
\alias{delete_user_from_view}
\title{функция для удаления пользователя из представления.}
\usage{
delete_user_from_view(acc_id, webproperty_id, view_id, email_to_delete)
}
\description{
функция для удаления пользователя из представления.
}
