% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delete_users_permissions.R
\name{delete_user_from_webproperty}
\alias{delete_user_from_webproperty}
\title{функция для удаления пользователя из ресурса.}
\usage{
delete_user_from_webproperty(acc_id, webproperty_id, email_to_delete)
}
\description{
функция для удаления пользователя из ресурса.
}
