% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delete_users_permissions.R
\name{delete_users_permissions}
\alias{delete_users_permissions}
\title{перебор аккаунтов, ресурсов и представлений с вызовом функций удаления на каждом шаге}
\usage{
delete_users_permissions(emails_to_delete)
}
\description{
перебор аккаунтов, ресурсов и представлений с вызовом функций удаления на каждом шаге
}
