% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delete_users_permissions.R
\name{inspect_permissions}
\alias{inspect_permissions}
\title{функция для проверки наличия необходимых прав доступа для просмотра и удаления пользователей аккаунта}
\usage{
inspect_permissions(x, n, ga_level)
}
\description{
функция для проверки наличия необходимых прав доступа для просмотра и удаления пользователей аккаунта
}
