% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delete_users_permissions.R
\name{delete_users_permissions}
\alias{delete_users_permissions}
\title{Management API: Delete users permissions}
\usage{
delete_users_permissions(emails_to_delete)
}
\arguments{
\item{emails_to_delete}{List of emails}
}
\description{
function gets an emails list as an input and removes access for each email from Google Analytics accounts on every hierarchy level.
}
